士兵的移动、攻击、追踪基本好了，撤退行为需要重写，还有就是到边缘位置似乎会卡住，这个再看看

+++ Assets/Behavior Designer Tactical/Scripts/Demo/Bullet.cs
@@ -7,7 +7,7 @@ namespace BehaviorDesigner.Runtime.Tactical
         // The speed of the bullet
         public float speed;
         // The amount of damage the bullet does
-        public float damageAmount = 5;
+        public int damageAmount = 5;
         // Destroy itself after this amount of time
         public float selfDestructTime = 5;
